haku/state_template: bazelize the UI (Dockerfile+npm → Bazel+pnpm)
Port the validated haku-state Bazel build onto the state_template seed's
pruned UI, replacing the Dockerfile + npm pipeline with a Bazel + pnpm build
that mirrors haku-state's CI-green form:

- Root Bazel setup: MODULE.bazel (rules_python/aspect_rules_js/ts/py, rules_oci,
  pnpm 9.15.9, node 22), .bazelrc (local exec, no RBE, --disk_cache in the
  runner's home bazelrc), .bazelversion, .bazelignore, .gitattributes, BUILD.bazel.
- ui/BUILD.bazel: rules_oci image (backend py_image_layer + vite dist tar on
  debian-slim), oci_load/oci_push, build_test gate.
- ui/backend/BUILD.bazel: py_library per module, aspect py_binary server,
  py_test targets (no SSOT export_schema — the template's UI has no codegen).
- ui/frontend/BUILD.bazel: node_modules link, vite dist, tsc/eslint/vitest gates
  (no openapi-typescript schema genrule — pruned UI has no generated src/api).
- eslint runs in Bazel (eslint_test); pnpm migration: pnpm-lock.yaml generated in
  isolation, package-lock.json deleted, esbuild onlyBuiltDependencies.
- pytest_bazel.main() entry points on every backend test.
- tsconfig types:["vite/client"] replaces src/vite-env.d.ts (deleted); extract
  mdx link helpers to mdx_links.ts and drop dead widgets.tsx WIDGETS export so
  eslint react-refresh passes.
- Workflows: bazel-ci.yaml (no ci-logs bring-up scaffolding), ruff-only lint.yaml,
  validate-state.yaml via //tools:validate_state; delete Dockerfile + build-ui.yaml.
- Exclude the seed from ducktape's own build via root .bazelignore.

Validated in an isolated pnpm dir: tsc --noEmit, eslint ., vitest (16 files /
83 tests), and backend ruff check + format all clean.
