+++ haku/state_template/.bazelignore
@@ -1,0 +1,1 @@
+ui/frontend/node_modules
