+++ haku/state_template/.bazelrc
@@ -1,0 +1,55 @@
+common --enable_bzlmod
+
+# ── Local execution only (NO BuildBuddy / RBE) ───────────────────────────────
+# haku-state deliberately runs everything locally: it may hold private operator
+# info, so its source never leaves the cluster's own Forgejo Actions runner.
+# There is no --remote_executor / --remote_cache. An on-disk cache keeps rebuilds
+# fast, but --disk_cache is a machine-specific path so it lives in each user's /
+# the CI runner's ~/.bazelrc (a standard home bazelrc Bazel reads automatically),
+# not here — see tools/ci/trust_egress_ca.sh for the runner's.
+build --spawn_strategy=local
+
+# Strict action env: actions get a fixed PATH instead of inheriting the client's
+# PATH. Keeps builds reproducible across the runner and dev machines.
+build --incompatible_strict_action_env
+
+# Don't generate empty __init__.py stubs in runfiles (namespace-package hygiene).
+common --incompatible_default_to_explicit_init_py
+
+# Use the shell bootstrap so py_binary invokes the hermetic interpreter directly
+# rather than `#!/usr/bin/env python3` (avoids _sre.MAGIC mismatch / stdlib
+# shadowing when a system python differs from the 3.13 toolchain).
+common --@rules_python//python/config_settings:bootstrap_impl=script
+
+# Pin the Python version so mixed toolchains never leak in.
+build --@rules_python//python/config_settings:python_version=3.13
+
+# Quiet, non-interactive output for CI logs.
+common:ci --curses=no
+common:ci --show_progress_rate_limit=30
+common:ci --announce_rc
+# Print failing test logs inline (they're otherwise only in bazel-testlogs, which
+# CI doesn't upload) so a failure is diagnosable from the run output.
+test:ci --test_output=errors
+
+# ── Egress: proxy + CA for the in-cluster CI runner ──────────────────────────
+# The haku-ci runner reaches bcr.bazel.build / GitHub / nodejs.org / npm / PyPI
+# through haku-egress-proxy (a TLS-bumping forward proxy). Two needs, neither in
+# this checked-in .bazelrc so a local `bazel build` (direct egress) is unaffected:
+#
+#  - Proxy routing: the runner injects HTTP(S)_PROXY / NO_PROXY into the job
+#    container env (cluster/k8s/haku-ci/config.yaml). Bazel's repository
+#    downloader honors the http_proxy/https_proxy env vars, so no flag is needed.
+#  - CA trust: the proxy bumps TLS with the cluster's egress CA. Bazel's
+#    downloader runs in the server JVM, which validates against a Java KeyStore —
+#    NOT SSL_CERT_FILE/CURL_CA_BUNDLE (those cover curl/python/node). The CI job
+#    imports the CA into the distro Java truststore and points the server JVM at
+#    it via the runner's ~/.bazelrc (Bazel reads the home bazelrc automatically,
+#    so no checked-in try-import is needed and local/CLI builds are unaffected) —
+#    see .forgejo/workflows/bazel-ci.yaml ("Trust egress-proxy CA") and
+#    tools/ci/trust_egress_ca.sh.
+#
+# TODO(egress-transparency): make CA trust less bespoke than a per-run keytool
+#   import into ~/.bazelrc — e.g. JAVA_TOOL_OPTIONS from the runner env, or bake
+#   the CA into the job image / point Bazel at a system JDK whose cacerts
+#   ca-certificates-java already syncs. (operator, this session)
