+++ haku/state_template/.bazelversion
@@ -1,0 +1,1 @@
+8.6.0
