--- haku/state_template/.forgejo/workflows/build-ui.yaml
@@ -1,91 +1,0 @@
-# Forgejo Actions workflow that tests, builds, and ships Haku's UI image.
-#
-# Runs IN-CLUSTER on the contained, repo-scoped Forgejo Actions runner (so haku-state's
-# source never leaves the cluster — it may hold private state) — NOT on any external CI.
-# The runner is rootless and runs jobs in a standard "act" image (node + git + docker CLI),
-# wired to a dind sidecar so `docker build` works.
-#
-# Two jobs, gated:
-#   test  — `docker build --target test` runs the backend pytest suite, a frontend build
-#           that runs the tsc typecheck, and `--target frontend-test` (vitest). If anything
-#           fails, the build fails.
-#   build — `needs: test`, so it only runs once tests pass. Builds the runtime image
-#           and pushes it to the Forgejo registry as
-#           forgejo.example.com/haku/ui:main-<utc-timestamp>-<short-sha>.
-# So a red test suite blocks the push and tag entirely — no image ships untested.
-#
-# CI does NOT touch the deployment manifest: Flux image automation watches the
-# registry, picks the newest tag by its timestamp, and writes it into the deployment
-# itself (the `{"$imagepolicy": ...}` marker). So there is no CI commit-back and no
-# `[skip ci]` loop.
-#
-# This is STARTER source Haku adopts and evolves. Heavily commented on purpose.
-# Fill in `haku-ci` with the label of your in-cluster Forgejo Actions runner.
-
-name: build-ui
-on:
-  push:
-    branches: [main]
-    # Only re-test/rebuild when the UI source or this workflow changes.
-    paths:
-      - "ui/**"
-      - ".forgejo/workflows/build-ui.yaml"
-
-jobs:
-  # Gate: run the tests before anything is built for real or pushed. The test stage
-  # installs dev deps and runs pytest during `docker build`; a non-zero exit fails the
-  # build, fails this job, and (via `needs: test` below) stops the push.
-  test:
-    runs-on: haku-ci
-    permissions:
-      contents: read
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      # Backend unit tests: the test stage's `RUN python -m pytest -q` runs them here.
-      # `--target test` stops at that stage (no SPA/runtime layers built), so it's fast.
-      - name: Backend tests
-        run: docker build -f ui/Dockerfile --target test -t haku-ui-test ui
-
-      # Frontend typecheck + build: the frontend stage runs `tsc -b && vite build`, so
-      # a type error fails the build. Building the `frontend` target exercises that
-      # without producing a pushable image (cached, so the real build reuses it).
-      - name: Frontend typecheck and build
-        run: docker build -f ui/Dockerfile --target frontend -t haku-ui-frontend ui
-
-      # Frontend unit tests: the frontend-test stage's `RUN npm run test` (vitest) runs here.
-      # Reuses the cached frontend layer, so it only adds the test execution.
-      - name: Frontend tests
-        run: docker build -f ui/Dockerfile --target frontend-test -t haku-ui-frontend-test ui
-
-  build:
-    needs: test
-    runs-on: haku-ci # the repo-scoped, contained Forgejo Actions runner label
-    permissions:
-      contents: read
-      packages: write
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      # Auth to the Forgejo registry over HTTPS (`forgejo.example.com`, valid cert).
-      # NOT the built-in job token: Forgejo's auto github.token cannot push packages
-      # (the `permissions: packages: write` block isn't honored yet — forgejo#3571,
-      # targeted for Forgejo 16). So we log in as the repo owner (`haku`) with the
-      # REGISTRY_PUSH_TOKEN repo Action secret you provision for this repo.
-      - name: Log in to the Forgejo registry
-        run: |
-          echo "${{ secrets.REGISTRY_PUSH_TOKEN }}" \
-            | docker login forgejo.example.com -u "${{ github.repository_owner }}" --password-stdin
-
-      # Tag pattern is what Flux image automation sorts on: a UTC timestamp makes
-      # the newest build the highest tag (ImagePolicy `numerical` on the captured
-      # timestamp), and the short sha keeps it traceable to a commit. The runtime
-      # image reuses the layers the test job already built (dind layer cache).
-      - name: Build and push the image
-        run: |
-          TAG="main-$(date -u +%Y%m%d%H%M%S)-$(echo "${{ github.sha }}" | cut -c1-7)"
-          IMAGE="forgejo.example.com/haku/ui:${TAG}"
-          docker build -f ui/Dockerfile --target runtime --build-arg GIT_SHA="${{ github.sha }}" -t "$IMAGE" ui
-          docker push "$IMAGE"
