+++ haku/state_template/.forgejo/workflows/lint.yaml
@@ -1,0 +1,53 @@
+# Backend ruff lint gate, run as a standalone CI step rather than a Bazel target:
+# ruff is a fast self-contained binary that gains nothing from Bazel's graph, and
+# its wheel has no console_scripts entry point (awkward to wrap). Frontend eslint,
+# by contrast, IS a Bazel target (//ui/frontend:eslint_test) — it reuses the
+# node_modules rules_js already links, avoiding a separate pnpm install that the
+# squid egress proxy couldn't serve reliably (pnpm's parallel fetch timed out).
+# tsc + vitest + eslint + the backend pytest suite + the image build run in
+# bazel-ci.yaml; only ruff is here.
+#
+# Bazel-free, via `docker build` on the dind sidecar — the same pattern as
+# validate-state.yaml (a `docker run -v` bind mount would resolve on the sidecar's
+# filesystem, not the job container's checkout; building streams the repo as the
+# context). A failing RUN fails the build, which fails the job.
+
+name: lint
+on:
+  push:
+    branches: [main]
+    paths:
+      - "ui/backend/**"
+      - ".forgejo/workflows/lint.yaml"
+
+jobs:
+  lint:
+    runs-on: haku-ci
+    permissions:
+      contents: read
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      # CA + proxy gotcha (found & fixed 2026-07-04, see
+      # items/validate-state-ci-docker-build-failing-2026-07.md): docker build's RUN
+      # steps don't reliably inherit this job container's HTTP_PROXY/HTTPS_PROXY, and
+      # pip ignores the system CA store — it needs the proxy forwarded via --build-arg
+      # and the egress CA pointed at explicitly (PIP_CERT). Stage the runner-mounted CA
+      # into the build context; the build COPYs it in before pip hits pypi.
+      - name: Prepare the egress-proxy CA for the build
+        run: cp /egress-proxy-ca/ca-certificates.crt ci-mitm-ca.crt
+
+      # ruff auto-discovers [tool.ruff] in ui/backend/pyproject.toml. Keep the pin
+      # in sync with ui/backend/pyproject.toml / requirements.txt.
+      - name: ruff (backend)
+        run: |
+          docker build --build-arg HTTP_PROXY="$HTTP_PROXY" --build-arg HTTPS_PROXY="$HTTPS_PROXY" -f - . <<'EOF'
+          FROM python:3.13-slim
+          COPY ci-mitm-ca.crt /usr/local/share/ca-certificates/egress-proxy.crt
+          ENV PIP_CERT=/usr/local/share/ca-certificates/egress-proxy.crt
+          RUN pip install --quiet --no-cache-dir ruff==0.15.20
+          WORKDIR /repo
+          COPY . .
+          RUN ruff check ui/backend && ruff format --check ui/backend
+          EOF
