+++ haku/state_template/.gitignore
@@ -1,0 +1,12 @@
+ui/frontend/dist/
+ui/frontend/node_modules/
+__pycache__/
+*.pyc
+
+# Bazel: convenience symlinks + local overrides. Build outputs live in the
+# --disk_cache (~/.cache/bazel-disk) and the sandboxed output base, not in-tree.
+/bazel-*
+
+# Egress-proxy CA, staged into the docker-build context by .forgejo/workflows/lint.yaml
+# at CI time (copied from the runner mount). Never committed.
+/ci-mitm-ca.crt
