+++ haku/state_template/tools/BUILD.bazel
@@ -1,0 +1,17 @@
+load("@rules_python//python:defs.bzl", "py_binary")
+
+# State validator: parses every state file the UI backend reads (items/, kitchen/,
+# runs/, responses/, memory/improvements/) against ui/backend/models.py — the real
+# read contract. Run over the working tree:  bazel run //tools:validate_state
+# (the script honors BUILD_WORKSPACE_DIRECTORY; see validate_state.py).
+py_binary(
+    name = "validate_state",
+    srcs = ["validate_state.py"],
+    main = "validate_state.py",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//ui/backend:models",
+        "@pypi//pydantic",
+        "@pypi//pyyaml",
+    ],
+)
