+++ haku/state_template/ui/frontend/.gitignore
@@ -1,5 +1,5 @@
-# Build output and deps — never committed. The build artifact is the CI-built
-# container image (see ../Dockerfile, ../../.forgejo/workflows/build-ui.yaml).
+# Build output and deps — never committed. The build artifact is the Bazel-built
+# container image (//ui:image, pushed by ../../.forgejo/workflows/bazel-ci.yaml).
 node_modules/
 dist/
 *.tsbuildinfo
