+++ haku/state_template/ui/frontend/tsconfig.json
@@ -16,7 +16,14 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+    // Load vite/client's ambients (incl. `declare module "*.css"`, which the css
+    // side-effect imports in main.tsx need under noUncheckedSideEffectImports) from
+    // node_modules directly. The conventional Vite `vite-env.d.ts` triple-slash
+    // reference can't do this under Bazel: js_library routes a source .d.ts to the
+    // typings output group, so the tsc_test sandbox never receives it — but it does
+    // receive node_modules (:node_modules), so `types` resolves it every time.
+    "types": ["vite/client"]
   },
   "include": ["src"]
 }
